+++ changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 
+- Mark current day by default in gantt chart
 - NewGanttChartFrom constructor
 
 ### Changed